+++ stroom-authentication/stroom-authentication-api/src/main/java/stroom/authentication/api/OpenIdClientDetailsFactory.java
@@ -1,0 +1,12 @@
+package stroom.authentication.api;
+
+public interface OpenIdClientDetailsFactory {
+//    String getClientId();
+//
+//    String getClientSecret();
+
+    OAuth2Client getOAuth2Client();
+
+    OAuth2Client getOAuth2Client(final String clientId);
+
+}