+++ stroom-authentication/stroom-authentication-common/src/main/java/stroom/authentication/config/AuthenticationConfig.java
@@ -19,68 +19,89 @@
 package stroom.authentication.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import stroom.util.cache.CacheConfig;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import stroom.util.config.annotations.ReadOnly;
+import stroom.util.config.annotations.RequiresRestart;
 import stroom.util.shared.AbstractConfig;
-import stroom.util.time.StroomDuration;
+import stroom.util.shared.validation.ValidationSeverity;
 
 import javax.annotation.Nullable;
-import javax.validation.Valid;
+import javax.inject.Singleton;
+import javax.validation.constraints.AssertFalse;
 import javax.validation.constraints.NotNull;
 
+@Singleton
+@JsonPropertyOrder(alphabetic = true)
 public final class AuthenticationConfig extends AbstractConfig {
+
     public static final String PROP_NAME_EMAIL = "email";
     public static final String PROP_NAME_TOKEN = "token";
     public static final String PROP_NAME_OAUTH2 = "oauth2";
     public static final String PROP_NAME_PASSWORD_INTEGRITY_CHECKS = "passwordIntegrityChecks";
 
-    @Nullable
-    @JsonProperty(PROP_NAME_EMAIL)
-    public EmailConfig emailConfig = new EmailConfig();
-    @Valid
-    @NotNull
-    @JsonProperty
+    private boolean useDefaultOpenIdCredentials = true;
     private String certificateDnPattern = ".*\\((.*)\\)";
-    @Valid
-    @NotNull
-    @JsonProperty
     private int certificateDnCaptureGroupIndex = 1;
-    @Valid
-    @NotNull
-    @JsonProperty
     private String loginUrl = "/s/login";
-    @Valid
-    @NotNull
-    @JsonProperty
     private String changePasswordUrl = "/s/changepassword";
-    @Nullable
-    @JsonProperty
     private Integer failedLoginLockThreshold = 3;
-    @Nullable
-    @JsonProperty
     private String unauthorisedUrl = "/s/unauthorised";
-    @Nullable
-    @JsonProperty(PROP_NAME_TOKEN)
-    private TokenConfig tokenConfig = new TokenConfig();
-    @NotNull
-    @JsonProperty(PROP_NAME_OAUTH2)
-    private OAuth2Config oAuth2Config = new OAuth2Config();
 
-    @NotNull
-    @JsonProperty(PROP_NAME_PASSWORD_INTEGRITY_CHECKS)
+    private EmailConfig emailConfig = new EmailConfig();
+    private TokenConfig tokenConfig = new TokenConfig();
+    private OAuth2Config oAuth2Config = new OAuth2Config();
     private PasswordIntegrityChecksConfig passwordIntegrityChecksConfig = new PasswordIntegrityChecksConfig();
 
+    @AssertFalse(
+            message = "Using default OpenId authentication credentials. These should only be used " +
+                    "in test/demo environments. Set stroom.authentication.useDefaultOpenIdCredentials to false for " +
+                    "production environments.",
+            payload = ValidationSeverity.Warning.class)
+    @RequiresRestart(RequiresRestart.RestartScope.SYSTEM)
+    @ReadOnly
+    @JsonProperty()
+    @JsonPropertyDescription("If true, stroom will use a set of default authentication credentials to allow" +
+            "API calls from stroom-proxy. For test or demonstration purposes only, set to false for production")
+    public boolean isUseDefaultOpenIdCredentials() {
+        return useDefaultOpenIdCredentials;
+    }
+
+    public void setUseDefaultOpenIdCredentials(final boolean useDefaultOpenIdCredentials) {
+        this.useDefaultOpenIdCredentials = useDefaultOpenIdCredentials;
+    }
+
+    @NotNull
+    @JsonProperty
     public final String getCertificateDnPattern() {
         return this.certificateDnPattern;
     }
 
+    public void setCertificateDnPattern(final String certificateDnPattern) {
+        this.certificateDnPattern = certificateDnPattern;
+    }
+
+    @NotNull
+    @JsonProperty
     public final String getLoginUrl() {
         return this.loginUrl;
     }
 
+    public void setLoginUrl(final String loginUrl) {
+        this.loginUrl = loginUrl;
+    }
+
+    @NotNull
+    @JsonProperty
     public String getChangePasswordUrl() {
         return changePasswordUrl;
     }
 
+    public void setChangePasswordUrl(final String changePasswordUrl) {
+        this.changePasswordUrl = changePasswordUrl;
+    }
+
+    @Nullable
     @JsonProperty(PROP_NAME_EMAIL)
     public EmailConfig getEmailConfig() {
         return emailConfig;
@@ -90,10 +111,17 @@ public final class AuthenticationConfig extends AbstractConfig {
         this.emailConfig = emailConfig;
     }
 
+    @Nullable
+    @JsonProperty
     public Integer getFailedLoginLockThreshold() {
         return this.failedLoginLockThreshold;
     }
 
+    public void setFailedLoginLockThreshold(final Integer failedLoginLockThreshold) {
+        this.failedLoginLockThreshold = failedLoginLockThreshold;
+    }
+
+    @Nullable
     @JsonProperty(PROP_NAME_TOKEN)
     public TokenConfig getTokenConfig() {
         return tokenConfig;
@@ -103,6 +131,7 @@ public final class AuthenticationConfig extends AbstractConfig {
         this.tokenConfig = tokenConfig;
     }
 
+    @NotNull
     @JsonProperty(PROP_NAME_OAUTH2)
     public OAuth2Config getOAuth2Config() {
         return oAuth2Config;
@@ -113,10 +142,16 @@ public final class AuthenticationConfig extends AbstractConfig {
     }
 
     @Nullable
+    @JsonProperty
     public String getUnauthorisedUrl() {
         return unauthorisedUrl;
     }
 
+    public void setUnauthorisedUrl(final String unauthorisedUrl) {
+        this.unauthorisedUrl = unauthorisedUrl;
+    }
+
+    @NotNull
     @JsonProperty(PROP_NAME_PASSWORD_INTEGRITY_CHECKS)
     public PasswordIntegrityChecksConfig getPasswordIntegrityChecksConfig() {
         return passwordIntegrityChecksConfig;
@@ -126,7 +161,13 @@ public final class AuthenticationConfig extends AbstractConfig {
         this.passwordIntegrityChecksConfig = passwordIntegrityChecksConfig;
     }
 
+    @NotNull
+    @JsonProperty
     public int getCertificateDnCaptureGroupIndex() {
         return certificateDnCaptureGroupIndex;
+    }
+
+    public void setCertificateDnCaptureGroupIndex(final int certificateDnCaptureGroupIndex) {
+        this.certificateDnCaptureGroupIndex = certificateDnCaptureGroupIndex;
     }
 }