+++ stroom-security/stroom-security-impl/src/main/java/stroom/security/impl/OpenIdConfig.java
@@ -1,9 +1,11 @@
 package stroom.security.impl;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import stroom.util.shared.AbstractConfig;
 
 import javax.inject.Singleton;
+import javax.validation.constraints.NotNull;
 
 @Singleton
 public class OpenIdConfig extends AbstractConfig {
@@ -29,6 +31,9 @@ public class OpenIdConfig extends AbstractConfig {
      * @return true if Stroom will handle the OpenId authenication, false if an external
      * OpenId provider is used.
      */
+    @JsonProperty
+    @JsonPropertyDescription("True if Stroom will handle OpenId authentication, false if an " +
+            "external OpenId provider is to be used.")
     public boolean isUseInternal() {
         return useInternal;
     }
@@ -37,6 +42,10 @@ public class OpenIdConfig extends AbstractConfig {
         this.useInternal = useInternal;
     }
 
+    @NotNull
+    @JsonProperty
+    @JsonPropertyDescription("The issuer used in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getIssuer() {
         return issuer;
     }
@@ -45,6 +54,10 @@ public class OpenIdConfig extends AbstractConfig {
         this.issuer = issuer;
     }
 
+    @NotNull
+    @JsonProperty
+    @JsonPropertyDescription("The authentication endpoint used in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getAuthEndpoint() {
         return authEndpoint;
     }
@@ -53,6 +66,10 @@ public class OpenIdConfig extends AbstractConfig {
         this.authEndpoint = authEndpoint;
     }
 
+    @NotNull
+    @JsonProperty
+    @JsonPropertyDescription("The token endpoint used in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getTokenEndpoint() {
         return tokenEndpoint;
     }
@@ -61,6 +78,10 @@ public class OpenIdConfig extends AbstractConfig {
         this.tokenEndpoint = tokenEndpoint;
     }
 
+    @NotNull
+    @JsonProperty
+    @JsonPropertyDescription("The URI to obtain the JSON Web Key Set from in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getJwksUri() {
         return jwksUri;
     }
@@ -69,7 +90,11 @@ public class OpenIdConfig extends AbstractConfig {
         this.jwksUri = jwksUri;
     }
 
+    // TODO Not sure we can add NotNull to this as it has no default and if useInternal is true
+    //  it doesn't need a value
     @JsonProperty(PROP_NAME_CLIENT_ID)
+    @JsonPropertyDescription("The client ID used in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getClientId() {
         return clientId;
     }
@@ -78,7 +103,11 @@ public class OpenIdConfig extends AbstractConfig {
         this.clientId = clientId;
     }
 
+    // TODO Not sure we can add NotNull to this as it has no default and if useInternal is true
+    //  it doesn't need a value
     @JsonProperty(PROP_NAME_CLIENT_SECRET)
+    @JsonPropertyDescription("The client secret used in OpenId authentication." +
+            "Should only be set if useInternal is true.")
     public String getClientSecret() {
         return clientSecret;
     }